Validator Docs and Aliases
- Added aliases
- Added validator docs
- Restructured tests
