+++ src/Exception/CircularReferenceException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Krak\Validation\Exception;
+
+final class CircularReferenceException extends ValidationException
+{
+
+}
