+++ src/Exception/ValidationException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Krak\Validation\Exception;
+
+class ValidationException extends \Exception
+{
+
+}
