+++ test/unit/violation.spec.php
@@ -1,0 +1,16 @@
+<?php
+
+use Krak\Validation;
+
+describe('Violation', function() {
+    describe('->abort', function() {
+        it('throws a violation exception', function() {
+            try {
+                Validation\violate('code')->abort();
+                assert(false);
+            } catch (Validation\Exception\ViolationException $e) {
+                assert($e->violation->code == 'code');
+            };
+        });
+    });
+});
