--- test/violation.php
@@ -1,14 +1,0 @@
-<?php
-
-use Krak\Validation;
-
-describe('->abort', function() {
-    it('throws a violation exception', function() {
-        try {
-            Validation\violate('code')->abort();
-            assert(false);
-        } catch (Validation\Exception\ViolationException $e) {
-            assert($e->violation->code == 'code');
-        };
-    });
-});
